+++ src/main/Data/Scripts/HelmetOffDialog/RegisterOnTalkEvents.lua
@@ -4,10 +4,13 @@ function Entity.GetPos(...)
     local log = HelmetOffDialog:log()
     log:info("OnTalkEvent: Triggered by Entity.GetPos")
 
+    --- @type _G.Entity
+    local this = ...
+
     HelmetOffDialog:error():catch(function()
         --- @type OnTalkEvent
         local event = HelmetOffDialog:onTalkEvent()
-        event:handle()
+        event:handle(this)
     end)
 
     return original_Entity_GetPos(...)