+++ src/main/Data/Scripts/HelmetOffDialog/RegisterTradeMenuEvents.lua
@@ -1,0 +1,18 @@
+function HelmetOffDialog:ApseCharacterOnInit(
+        elementName, instanceId, eventName, argTable
+)
+    local log = HelmetOffDialog:log()
+    log:info("ApseCharacterOnInit", elementName, instanceId, eventName, argTable)
+
+    if eventName ~= "OnInit" then
+        return
+    end
+
+    --- @type TalkEndedEvent
+    local event = HelmetOffDialog:talkEndedEvent()
+    event:forcePutOn()
+end
+
+UIAction.RegisterElementListener(
+        HelmetOffDialog, "ApseCharacter", -1, "", "ApseCharacterOnInit"
+)