+++ src/main/Data/Scripts/Startup/HelmetOffDialog_Startup.lua
@@ -4,6 +4,7 @@ _G.HelmetOffDialog:init()
 
 function HelmetOffDialog:OnGameplayStartKcdOne()
     Script.LoadScript("Scripts/HelmetOffDialog/RegisterOnTalkEvents.lua")
+    Script.LoadScript("Scripts/HelmetOffDialog/RegisterTradeMenuEvents.lua")
 end
 
 UIAction.RegisterEventSystemListener(